[networking] Add module (#2861)
No actual changes - code is only moved around
